+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Initial release of the package.
+- Added a Beautiful, Customizable Glassmorphism Bottom Navigation Bar.